# booking/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: A Book API page
paths:
  /:
    get:
      summary: home
      description: A home page.
      tags:
        - Home
      responses:
        "200":
          description: A home reponse
  /order/{id}:
    get:
      summary: get order by id
      description: get order by id.
      tags:
        - Order
      parameters:
        - $ref: "#/components/parameters/orderId"
      responses:
        "200":
          description: A order reponse
    put:
      summary:  update order by id
      description:  update order by id.
      tags:
        - Order
      parameters:
        - $ref: "#/components/parameters/orderId"
      responses:
        "200":
          description: An order update response
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createOrderRequestBody"
    delete:
      summary:  delete order by id
      description:  delete order by id.
      tags:
        - Order
      parameters:
        - $ref: "#/components/parameters/orderId"
      responses:
        "200":
          description: A order deletion response
  /order:
    post:
      summary: create order
      description: Use this endpoint to create a order.
      tags:
        - Order
      responses:
        "201":
          description: Created
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createOrderRequestBody"
  /order/{page}:/{pageSize}:
    get:
      summary: get orders
      description: get orders
      tags:
        - Order
      parameters:
        - $ref: "#/components/parameters/orderPage"
        - $ref: "#/components/parameters/orderPageSize"
      responses:
        "200":
          description: Orders response          
  /bulk/order:
    post:
      summary: create orders
      description: Use this endpoint to create orders.
      tags:
        - Bulk Order
      responses:
        "201":
          description: Created
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: "#/components/schemas/createOrderRequestBody"
    put:
      summary: update orders
      description: Use this endpoint to update orders.
      tags:
        - Bulk Order
      responses:
        "200":
          description: Updated
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: "#/components/schemas/updateOrderRequestBody"
  /bulk/order/{ids}:
    get:
      summary: get order by ids
      description: get order by ids.
      tags:
        - Bulk Order
      parameters:
        - $ref: "#/components/parameters/orderIds"
      responses:
        "200":
          description: A order reponse
    delete:
      summary: delete order by ids
      description: delete order by ids.
      tags:
        - Bulk Order
      parameters:
        - $ref: "#/components/parameters/orderIds"
      responses:
        "200":
          description: A order reponse
                  
components:
  parameters:
    orderId:
      in: path
      name: id
      description:  id of order
      required: true
      schema:
        type: string
    orderIds:
      in: path
      name: ids
      description:  ids of orders
      required: true
      schema:
        type: array
        items: 
          type: string
    orderPage:
      in: path
      name: page
      description:  page of orders
      required: true
      schema:
        type: number
    orderPageSize:
      in: path
      name: pageSize
      description:  pageSize of orders
      required: true
      schema:
        type: number
  schemas:
    updateOrderRequestBody:
      type: object
      properties:
        id: 
          type: string
        price: 
          type: number
        status:
          type: string
    createOrderRequestBody:
      type: object
      properties:
        price: 
          type: number
        status:
          type: string

          